requirements.txt
fastapi
uvicorn
pandas
Faker
python-multipart
google-cloud-bigquery
google-cloud-storage